% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equalize_means.r
\name{equalize_logmeans}
\alias{equalize_logmeans}
\title{Equalize Average Decision Times}
\usage{
equalize_logmeans(x, formula, ...)
}
\arguments{
\item{x}{A data.frame containing the to-be-equalized variables and condition variables.}

\item{formula}{A formula specifying the model to be fitted, i.e. the dependent variable and the condition variables.}

\item{...}{Additional arguments passed to the \code{brms::brm} function.}
}
\value{
The data.frame \code{x} with the specified variable's average decision times equalized across conditions.
}
\description{
This function fits a Bayesian shifted lognormal model to reaction times (or another dependent variable) across conditions.
It then calculates the log means for the dependent variable across unique combinations of the independent variables,
effectively equalizing the average decision times based on the model.
}
\examples{
\dontrun{
data(pilot)

equalized_rt <- equalize_logmeans(
  x = pilot
  , formula = rt ~ condition * correct
  , chains = 3
  , cores = 1
  , iter = 2000
  , warmup = 1000
  , verbose = FALSE
  , refresh = 0
)
}
}
