% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equalize_means.r
\name{equalize_means}
\alias{equalize_means}
\title{Equalize Means Across Conditions}
\usage{
equalize_means(x, formula)
}
\arguments{
\item{x}{A data.frame containing the to-be-equalized variables and condition variables.}

\item{formula}{A formula specifying the variables to be equalized and the condition variables.}
}
\value{
The data.frame \code{x} with the specified variables' means equalized across conditions.
}
\description{
This function adjusts the means of specified variables across different conditions.
It calculates the mean of each variable for each combination of conditions, then subtracts the overall mean of these means from each,
effectively euqalizing the means of the variables across these groups.
}
\examples{
data(pilot)

equalized_data <- equalize_means(
  x = pilot,
  formula = cbind(acc_activity, dlpfc_activity) ~ condition * correct
)

}
