% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_data.r
\name{read_raw_data}
\alias{read_raw_data}
\title{Read Raw Data}
\usage{
read_raw_data(x = "data-raw/exp/")
}
\arguments{
\item{x}{A character string specifying the directory where the .xlsx files are located. Default is "data-raw/exp/".}
}
\value{
A data frame containing the combined data from all .xlsx files in the specified directory.
}
\description{
This function reads all .xlsx files from a specified directory and combine them into a single data frame.
The column names of the data frame are cleaned using the janitor::make_clean_names function.
}
\examples{
\dontrun{
read_raw_data("data-raw/exp/")
}

}
