% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_variable.r
\name{shuffle_variable}
\alias{shuffle_variable}
\title{Shuffle Covariate Values in a Data Frame}
\usage{
shuffle_variable(x, formula)
}
\arguments{
\item{x}{A data.frame containing the variables to be shuffled.}

\item{formula}{A formula specifying one or more covariates to shuffle (left-hand side) and one or more grouping variables (right-hand side).
The covariates will be shuffled within the levels of the grouping variables.}
}
\value{
The data.frame \code{x} the specified covariates shuffled within the levels of the grouping variables.
}
\description{
This function shuffles the values of one or more covariates within a data frame based on a specified formula.
The formula determines which variables are shuffled and how they are grouped during the shuffling process.
The shuffling is done in a way that maintains the relationship between the shuffled variables and other variables in the data frame.
}
\examples{
data(pilot)
shuffled_pilot <- shuffle_variable(
  x = pilot,
  formula = acc_activity ~ condition * correct
)
}
